+++ src/types/guild.ts
@@ -1,7 +1,13 @@
 import type { Snowflake } from "discord.js";
 
 import type { TLanguage } from "./language";
-import type { Member } from "./member";
+
+export interface Member {
+  name: string;
+  id: Snowflake;
+  joinedAt: number | null;
+  totalTime: number;
+}
 
 export interface Guild {
   name: string;
