+++ src/types/index.ts
@@ -1,5 +1,4 @@
 export * from "./embed";
 export * from "./guild";
 export * from "./language";
-export * from "./member";
 export * from "./vc";
