--- src/types/member.ts
@@ -1,8 +1,0 @@
-import type { Snowflake } from "discord.js";
-
-export interface Member {
-  name: string;
-  id: Snowflake;
-  joinedAt: number | null;
-  totalTime: number;
-}
